refactor: split ToastContext into context/hook files, fix timer cleanup ref
Eliminates two ESLint warnings:
- react-refresh: move useToast and ToastContext object to toastContext.ts
  so ToastContext.tsx only exports components (ToastProvider)
- react-hooks/exhaustive-deps: capture timerRef.current in local variable
  before cleanup function to avoid stale ref access

Closes #208
